Update backend dependencies to Spring Boot 3. Rewrite security config to work with new Spring Security style.
