+++ backend/src/main/java/cz/klecansky/projectmanagement/budget/io/BudgetCategoryEntity.java
@@ -3,7 +3,8 @@ package cz.klecansky.projectmanagement.budget.io;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
-import javax.persistence.*;
+
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
