+++ backend/src/main/java/cz/klecansky/projectmanagement/budget/io/BudgetItemEntity.java
@@ -1,7 +1,7 @@
 package cz.klecansky.projectmanagement.budget.io;
 
 import java.util.UUID;
-import javax.persistence.*;
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
