+++ backend/src/main/java/cz/klecansky/projectmanagement/comment/ui/request/CommentRequest.java
@@ -1,7 +1,7 @@
 package cz.klecansky.projectmanagement.comment.ui.request;
 
 import java.time.Instant;
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
 @Data
