+++ backend/src/main/java/cz/klecansky/projectmanagement/comment/ui/response/CommentResponse.java
@@ -3,8 +3,8 @@ package cz.klecansky.projectmanagement.comment.ui.response;
 import cz.klecansky.projectmanagement.user.ui.response.UserResponse;
 import java.time.Instant;
 import java.util.UUID;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Size;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
 import lombok.Data;
 
 @Data
