+++ backend/src/main/java/cz/klecansky/projectmanagement/group/io/entity/GroupMemberEntity.java
@@ -2,7 +2,7 @@ package cz.klecansky.projectmanagement.group.io.entity;
 
 import cz.klecansky.projectmanagement.user.io.entity.UserEntity;
 import java.util.UUID;
-import javax.persistence.*;
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
