+++ backend/src/main/java/cz/klecansky/projectmanagement/project/ui/request/ProjectRequest.java
@@ -1,7 +1,7 @@
 package cz.klecansky.projectmanagement.project.ui.request;
 
 import java.time.Instant;
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
 @Data
