+++ backend/src/main/java/cz/klecansky/projectmanagement/result/io/ResultEntity.java
@@ -2,7 +2,7 @@ package cz.klecansky.projectmanagement.result.io;
 
 import cz.klecansky.projectmanagement.outcome.io.OutcomeEntity;
 import java.util.UUID;
-import javax.persistence.*;
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
