+++ backend/src/main/java/cz/klecansky/projectmanagement/schedule/io/RowEntity.java
@@ -3,7 +3,7 @@ package cz.klecansky.projectmanagement.schedule.io;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
-import javax.persistence.*;
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
