+++ backend/src/main/java/cz/klecansky/projectmanagement/schedule/io/TaskScheduleEntity.java
@@ -2,7 +2,7 @@ package cz.klecansky.projectmanagement.schedule.io;
 
 import java.time.Instant;
 import java.util.UUID;
-import javax.persistence.*;
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
