--- backend/src/main/java/cz/klecansky/projectmanagement/security/config/WebSecurityConfig.java
@@ -1,62 +1,0 @@
-package cz.klecansky.projectmanagement.security.config;
-
-import static cz.klecansky.projectmanagement.security.SecurityConstants.PUBLIC_URLS;
-
-import cz.klecansky.projectmanagement.security.jwt.JWTAccessDeniedHandler;
-import cz.klecansky.projectmanagement.security.jwt.JWTAuthenticationEntryPoint;
-import cz.klecansky.projectmanagement.security.jwt.JWTConfigurer;
-import cz.klecansky.projectmanagement.security.jwt.TokenProvider;
-import lombok.AccessLevel;
-import lombok.RequiredArgsConstructor;
-import lombok.experimental.FieldDefaults;
-import org.springframework.context.annotation.Bean;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-@EnableWebSecurity
-@EnableGlobalMethodSecurity(prePostEnabled = true, securedEnabled = true)
-@FieldDefaults(makeFinal = true, level = AccessLevel.PRIVATE)
-@RequiredArgsConstructor
-public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
-    TokenProvider tokenProvider;
-    JWTAccessDeniedHandler jwtAccessDeniedHandler;
-    JWTAuthenticationEntryPoint jwtAuthenticationEntryPoint;
-
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-
-    @Override
-    protected void configure(HttpSecurity httpSecurity) throws Exception {
-        httpSecurity
-                .csrf()
-                .disable()
-                .exceptionHandling()
-                .accessDeniedHandler(jwtAccessDeniedHandler)
-                .authenticationEntryPoint(jwtAuthenticationEntryPoint)
-                .and()
-                .sessionManagement()
-                .sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-                .and()
-                .authorizeRequests()
-                .antMatchers(PUBLIC_URLS)
-                .permitAll()
-                .anyRequest()
-                .authenticated()
-                .and()
-                .apply(new JWTConfigurer(tokenProvider));
-    }
-
-    @Override
-    @Bean
-    public AuthenticationManager authenticationManagerBean() throws Exception {
-        return super.authenticationManagerBean();
-    }
-}
