+++ backend/src/main/java/cz/klecansky/projectmanagement/security/jwt/JWTAuthenticationEntryPoint.java
@@ -1,8 +1,8 @@
 package cz.klecansky.projectmanagement.security.jwt;
 
 import java.io.IOException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.stereotype.Component;
