+++ backend/src/main/java/cz/klecansky/projectmanagement/storage/ui/EditFileNameRequest.java
@@ -1,6 +1,6 @@
 package cz.klecansky.projectmanagement.storage.ui;
 
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
 @Data
