+++ backend/src/main/java/cz/klecansky/projectmanagement/user/io/entity/UserEntity.java
@@ -8,7 +8,7 @@ import cz.klecansky.projectmanagement.user.shared.Role;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
-import javax.persistence.*;
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
