+++ backend/src/main/java/cz/klecansky/projectmanagement/user/io/entity/VerificationTokenEntity.java
@@ -2,7 +2,7 @@ package cz.klecansky.projectmanagement.user.io.entity;
 
 import java.time.Instant;
 import java.util.UUID;
-import javax.persistence.*;
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
