+++ backend/src/main/java/cz/klecansky/projectmanagement/user/service/EmailService.java
@@ -14,7 +14,7 @@ import cz.klecansky.projectmanagement.user.shared.UserMapper;
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
 import java.util.UUID;
-import javax.mail.internet.MimeMessage;
+import jakarta.mail.internet.MimeMessage;
 import lombok.AccessLevel;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
