+++ backend/src/main/java/cz/klecansky/projectmanagement/user/service/UserService.java
@@ -1,21 +1,14 @@
 package cz.klecansky.projectmanagement.user.service;
 
 import cz.klecansky.projectmanagement.security.UserPrincipal;
-import cz.klecansky.projectmanagement.security.jwt.JWTToken;
-import cz.klecansky.projectmanagement.user.shared.NewUserPasscodeCommand;
 import cz.klecansky.projectmanagement.user.shared.UserCommand;
+import org.springframework.security.core.userdetails.UserDetailsService;
+
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
-import org.springframework.security.core.userdetails.UserDetailsService;
 
 public interface UserService extends UserDetailsService {
-    JWTToken signIn(String username, String password);
-
-    void forgottenPasswordRequest(String username);
-
-    boolean forgottenPasswordTokenCheck(UUID token);
-
     UserCommand createUser(UserCommand userCommand, String passcode);
 
     Optional<UserCommand> getUser(UUID userId);
@@ -28,9 +21,8 @@ public interface UserService extends UserDetailsService {
 
     List<UserCommand> getUsers();
 
+
     UserCommand newPassword(UUID passwordResetToken, String newPassword);
 
     UserCommand changePassword(UserPrincipal user, String currentPassword, String newPassword);
-
-    NewUserPasscodeCommand generateNewUserPasscode();
 }
