+++ backend/src/main/java/cz/klecansky/projectmanagement/user/ui/request/NewPasswordRequest.java
@@ -1,7 +1,7 @@
 package cz.klecansky.projectmanagement.user.ui.request;
 
 import java.util.UUID;
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
 
 @Data
