+++ backend/src/main/java/cz/klecansky/projectmanagement/user/ui/request/PasswordResetRequestRequest.java
@@ -1,7 +1,7 @@
 package cz.klecansky.projectmanagement.user.ui.request;
 
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
 
 @Data
